# rest-api-laravel56/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Todo API",
        "description": "Generated by l5-swagger",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": true,
                        "type": "string",
                        "format": "email"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "c_password",
                        "in": "formData",
                        "description": "Confirm Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user data with access_token",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Joao"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "12k091k2091283102k3.12093901283"
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "The username is the email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user data with access_token",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Joao"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "12k091k2091283102k3.12093901283"
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}